ref: #1170