@jfontan 2019-07-30
I've tested with several repositories and time is in the same ballpark as before but somehow uses more memory or is a but slower. This is a similar benchmark as the one provided but downloading a linux repo I have locally and downloading all history:

Before:

```
BenchmarkBareClone-4   	       1	202099566609 ns/op	87716304312 B/op	116607058 allocs/op
PASS
ok  	gopkg.in/src-d/go-git.v4	202.383s
```

After:

```
BenchmarkBareClone-4   	       1	204590089037 ns/op	95760605528 B/op	105199786 allocs/op
PASS
ok  	gopkg.in/src-d/go-git.v4	204.828s
```

I've also timed the execution of `_examples/clone` with the linux repository:

Before:

```
263.33user 11.80system 3:28.19elapsed 132%CPU (0avgtext+0avgdata 3009184maxresident)k
232inputs+3482592outputs (0major+812125minor)pagefaults 0swaps
```

After:

```
287.45user 13.48system 3:50.08elapsed 130%CPU (0avgtext+0avgdata 3073260maxresident)k
562240inputs+3482592outputs (0major+828890minor)pagefaults 0swaps
```

I don't know what can be happening as the changes look seem to improve memory allocation.