@orisano 2019-07-31
Please give me your memory profile.
```sh
go test -bench BareClone -benchmem -timeout 10000s -memprofile=mem.pb.gz -cpuprofile=cpu.pb.gz
```

I can't reproduce ...