@jfontan 2019-07-31
@orisano After your rebase I get similar numbers. These are the results of the tests with linux:

```
BenchmarkBareClone-4   	       1	201592414125 ns/op	95829895744 B/op	105199544 allocs/op
PASS
ok  	gopkg.in/src-d/go-git.v4	201.815s



271.64user 11.90system 3:28.21elapsed 136%CPU (0avgtext+0avgdata 2936412maxresident)k
0inputs+3482592outputs (0major+794100minor)pagefaults 0swaps
```

Bear in mind that this is not your benchmark. In your benchmark you are cloning with `Depth: 1`, this is cloning full history.

Thanks!

@mcuadros I've tested with smaller repos and I don't see any regression. Gains are not spectacular but I believe it's a good base to improve memory allocations.